requirements.txt
django
djangorestframework
psycopg2
python-dotenv
boto3